requirements.txt
click==8.0.3
colorama==0.4.4
dominate==2.6.0
et-xmlfile==1.1.0
Flask==2.0.2
Flask-Bootstrap4==4.0.2
itsdangerous==2.0.1
Jinja2==3.0.3
joblib==1.1.0
MarkupSafe==2.0.1
numpy==1.22.1
openpyxl==3.0.9
pandas==1.4.0
python-dateutil==2.8.2
pytz==2021.3
scikit-learn==1.0.2
scipy==1.7.3
six==1.16.0
sklearn==0.0
threadpoolctl==3.1.0
visitor==0.1.3
Werkzeug==2.0.2
XlsxWriter==3.0.2
